Fix the incorrect frequency being returned.
The problem was that my sampling frequency (8192 Hz) was not a multiple
of the clock rate (44100 Hz). This caused my fundamental frequency to
appear to be approximately 38% smaller than it should have been. The
note two and half steps below a given note is 2^(5/12) == 33% which
explains why I was seeing that frequency.